requirements.txt
temporalio>=1.7.0
requests>=2.31.0
strands-agents
boto3>=1.34.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
docker>=7.0.0
hypothesis>=6.0.0
